requirements.txt
alembic==1.14.1
email-validator==2.2.0
fastapi==0.115.12
uvicorn[standard]==0.34.3
sqlmodel==0.0.24
pydantic-settings==2.9.1
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
google-generativeai==0.8.4
python-dotenv==1.0.1
slowapi==0.1.9
redis==5.2.1
cachetools==5.3.3
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.0
